Bugfixes for odd behaviour
1) The receptors are positioned relative (getGlobalPosition) but I was placing
them absolutely in the svg() function. Relative is better as I don't need to
move them when the block is moved.
2) The dragBlocks of a BlockLinear sequence does not need to return all items
in the list, but only the next block in the chain. It recursively goes through
"children" and linear sequences are effectively child elements of the first
block in the chain.